requirements.txt
asgiref==3.10.0
attrs==25.4.0
bandit==1.7.9
black==24.4.2
click==8.3.0
Django==5.0.4
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.11
exceptiongroup==1.3.0
gunicorn==22.0.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
mdurl==0.1.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
psycopg2-binary==2.9.9
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.3.2
pytest-django==4.8.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rich==14.2.0
rpds-py==0.28.0
ruff==0.5.4
sqlparse==0.5.3
stevedore==5.5.0
tomli==2.3.0
typing_extensions==4.15.0
uritemplate==4.2.0
whitenoise==6.6.0
